basnetwrapper/requirements.txt
scikit_image==0.16.2
matplotlib==3.1.1
torch==1.5.0
Django==3.0.2
torchvision==0.6.0
numpy==1.16.4
djangorestframework==3.11.0
Pillow==7.1.2
python-decouple==3.3
skimage==0.0
